chore: refactor utils, revert assistant ID
