+++ src/app-config.json
@@ -2,7 +2,7 @@
   "accessibility": {
     "keyboardShortcut": "ctrl+?"
   },
-  "assistantId": "asst_v9QCHtVvhFqI5lmLjIZnSQUG",
+  "assistantId": "asst_xmAX5oxByssXrkBymMbcsVEm",
   "dimensions": {
     "height": 680,
     "width": 380
